Add Rails 5 support
This makes various changes to add support for Rails 5.

- Convert all `alias_method_chain` calls to `prepend` and `super` (`alias_method_chain` was removed in Rails 5).
- Move `QueryMethods` extensions to `Relation` (after dropping `alias_method_chain`, `jit_preload?` and other methods were no longer available from `AR::Relation`).
