+++ jit_preloader.gemspec
@@ -18,7 +18,7 @@ Gem::Specification.new do |spec|
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
 
-  spec.add_dependency "activerecord", "~> 4.2"
+  spec.add_dependency "activerecord", "> 4.2", "< 6"
   spec.add_dependency "activesupport"
 
   spec.add_development_dependency "bundler", "~> 1.7"
